Split has and curve from keypair
